split out index
this is the minimum necessary changes that passes all tests and checks
